Change how `TracePredictive` works, so it resets plates on resampling
Also use its own marginal, instead of the posteriors
